+++ unstructured/ingest/connector/fsspec.py
@@ -136,7 +136,7 @@ class FsspecIngestDoc(BaseIngestDoc):
 
     def cleanup_file(self):
         """Removes the local copy of the file after successful processing."""
-        if not self.config.preserve_downloads:
+        if not self.config.preserve_downloads and not self.config.download_only:
             logger.debug(f"Cleaning up {self}")
             os.unlink(self._tmp_download_file())
 
@@ -156,7 +156,7 @@ class FsspecConnector(BaseConnector):
         self.fs: AbstractFileSystem = get_filesystem_class(self.config.protocol)(
             **self.config.access_kwargs,
         )
-        self.cleanup_files = not config.preserve_downloads
+        self.cleanup_files = not config.preserve_downloads and not config.download_only
 
     def cleanup(self, cur_dir=None):
         """cleanup linginering empty sub-dirs from s3 paths, but leave remaining files
